[Refactor] 팀 선택 Template => Page 로직 이동
